Move module :faphub:installation to :faphub:installation:button
