+++ components/faphub/catalogtab/impl/build.gradle.kts
@@ -19,7 +19,7 @@ dependencies {
 
     implementation(projects.components.faphub.dao.api)
     implementation(projects.components.faphub.appcard.composable)
-    implementation(projects.components.faphub.installation.api)
+    implementation(projects.components.faphub.installation.button.api)
 
     // Compose
     implementation(libs.compose.ui)
