+++ components/faphub/category/impl/src/main/java/com/flipperdevices/faphub/category/impl/api/FapHubCategoryApiImpl.kt
@@ -11,7 +11,7 @@ import com.flipperdevices.faphub.category.api.FapHubCategoryApi
 import com.flipperdevices.faphub.category.impl.composable.ComposableFapHubCategory
 import com.flipperdevices.faphub.dao.api.model.FapCategory
 import com.flipperdevices.faphub.fapscreen.api.FapScreenApi
-import com.flipperdevices.faphub.installation.api.FapInstallationUIApi
+import com.flipperdevices.faphub.installation.button.api.FapInstallationUIApi
 import com.flipperdevices.faphub.search.api.FapHubSearchEntryApi
 import com.squareup.anvil.annotations.ContributesBinding
 import com.squareup.anvil.annotations.ContributesMultibinding
