+++ components/faphub/fapscreen/impl/build.gradle.kts
@@ -30,7 +30,7 @@ dependencies {
     implementation(projects.components.faphub.search.api)
     implementation(projects.components.faphub.dao.api)
     implementation(projects.components.faphub.appcard.composable)
-    implementation(projects.components.faphub.installation.api)
+    implementation(projects.components.faphub.installation.button.api)
 
     implementation(libs.kotlin.immutable.collections)
