+++ components/faphub/installation/button/impl/src/main/java/com/flipperdevices/faphub/installation/button/impl/model/FapInstallationState.kt
@@ -1,4 +1,4 @@
-package com.flipperdevices.faphub.installation.impl.model
+package com.flipperdevices.faphub.installation.button.impl.model
 
 import androidx.annotation.FloatRange
