+++ components/faphub/search/impl/src/main/java/com/flipperdevices/faphub/search/impl/api/FapHubSearchEntryApiImpl.kt
@@ -8,7 +8,7 @@ import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.di.provideDelegate
 import com.flipperdevices.core.ui.navigation.AggregateFeatureEntry
 import com.flipperdevices.faphub.fapscreen.api.FapScreenApi
-import com.flipperdevices.faphub.installation.api.FapInstallationUIApi
+import com.flipperdevices.faphub.installation.button.api.FapInstallationUIApi
 import com.flipperdevices.faphub.search.api.FapHubSearchEntryApi
 import com.flipperdevices.faphub.search.impl.composable.ComposableSearchScreen
 import com.squareup.anvil.annotations.ContributesBinding
