+++ instances/app/build.gradle.kts
@@ -162,8 +162,8 @@ dependencies {
     implementation(projects.components.faphub.fapscreen.api)
     implementation(projects.components.faphub.fapscreen.impl)
 
-    implementation(projects.components.faphub.installation.api)
-    implementation(projects.components.faphub.installation.impl)
+    implementation(projects.components.faphub.installation.button.api)
+    implementation(projects.components.faphub.installation.button.impl)
 
     implementation(projects.components.faphub.installedtab.api)
     implementation(projects.components.faphub.installedtab.impl)
